core polymorphic decoder
