+++ Tests/JSONTests/Utilities.swift
@@ -1,3 +1,4 @@
+import Core
 import JSON
 
 struct Name: JSONCodable {
